feat(study): unifica Estudiar en cursos multi-tema y elimina el flujo un-tema
Todo lo nuevo crea StudyPlan (acepta desde 1 tema): formulario
multi-tema embebido en Estudiar con ?courseId=, lista única de cursos
y fuera StudyUnitPage/StudyPlanCreatePage. Se eliminan el módulo API
study, los endpoints de generación un-tema sin consumidores y el
modelo StudyUnit con migración destructiva (borra los datos un-tema;
decisión de producto 09/07).
