+++ packages/shared/src/types/study.types.ts
@@ -1,9 +1,9 @@
 import type { TheoryModuleWithLessons } from './theory.types';
 
-// Dificultad de la unidad de estudio. Gobierna la generación de ejercicios y examen.
+// Dificultad de generación. Gobierna la generación de ejercicios y examen.
 export type StudyDifficulty = 'EASY' | 'MEDIUM' | 'HARD';
 
-// Ejercicio persistido en la unidad (misma forma que genera ExercisesService).
+// Ejercicio persistido en el plan (misma forma que genera ExercisesService).
 export type StudyExerciseType = 'SINGLE' | 'TRUE_FALSE' | 'OPEN';
 
 export interface StudyExercise {
@@ -14,34 +14,6 @@ export interface StudyExercise {
   explanation: string;
 }
 
-// Examen de la unidad (serializado SIN isCorrect, como getBank).
-export interface StudyExamAnswer {
-  id: string;
-  text: string;
-  order: number;
-}
-
-export interface StudyExamQuestion {
-  id: string;
-  text: string;
-  type: 'SINGLE' | 'MULTIPLE' | 'TRUE_FALSE';
-  order: number;
-  /// Tema del plan multi-tema al que pertenece la pregunta. Null/ausente en bancos un-tema.
-  topicLabel?: string | null;
-  answers: StudyExamAnswer[];
-}
-
-export interface StudyExam {
-  id: string;
-  title: string;
-  topic: string;
-  numQuestions: number;
-  timeLimit: number | null;
-  onlyOnce: boolean;
-  attemptCount: number;
-  questions: StudyExamQuestion[];
-}
-
 // Qué secciones se generaron con éxito.
 export interface StudySections {
   theory: boolean;
@@ -49,46 +21,9 @@ export interface StudySections {
   exam: boolean;
 }
 
-export interface StudyUnitSummary {
-  id: string;
-  topic: string;
-  title: string;
-  summary: string;
-  createdAt: string;
-  course: { id: string; title: string };
-  sections: StudySections;
-}
-
-export interface StudyUnitDetail extends StudyUnitSummary {
-  theory: TheoryModuleWithLessons | null;
-  exercises: StudyExercise[] | null;
-  exam: StudyExam | null;
-}
-
-// ── Requests ──
-export interface CreateStudyUnitRequest {
-  courseId: string;
-  topic: string;
-  numExercises: number;
-  difficulty: StudyDifficulty; // aplica a ejercicios y examen
-  numQuestions: 5 | 10;
-  timeLimit?: number; // segundos
-  onlyOnce?: boolean;
-}
-
-export interface RegenerateExercisesRequest {
-  count?: number;
-}
-
-export interface RegenerateExamRequest {
-  numQuestions?: 5 | 10;
-  timeLimit?: number;
-  onlyOnce?: boolean;
-}
-
-// ── Plan de estudio multi-tema (StudyPlan) ──
-// Combina N temas (oficiales del temario o libres) en teoría por tema +
-// ejercicios y examen combinados. Flujo adicional al un-tema (StudyUnit).
+// ── Plan de estudio (StudyPlan) ──
+// Combina 1..N temas (oficiales del temario o libres) en teoría por tema +
+// ejercicios y examen combinados. Es el flujo único de estudio.
 
 export type StudyTopicSource = 'OFFICIAL' | 'CUSTOM';
